Rename low end module for moment integrals
Seems like the name of the low level module (`moment_int`) will conflict with
the higher level module
